+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Task dependencies are declared properly, so Kotlin and similar tasks run after Jextract.
+
 ### Added
 
 - New implementations of both Jextract tasks have been added that process one library per task instance. This facilitates their usage and debugging problems with specific libraries.
+
+### Changed
+
+- If you configure the plugin using its DSL extension in your build script, nothing will change for you.
+- The plugin has been migrated to use the new tasks without worker API. If you configure the tasks directly in your build, this version will be a breaking change for you. Please try to configure the plugin via DSL, or migrate to the new tasks.
 
 ## [0.5.0] - 2025-05-18